feat(dev): remove future flags for CSS bundling
