+++ .changeset/remove-unstable-tailwind.md
@@ -5,4 +5,4 @@
 "@remix-run/testing": major
 ---
 
-Remove `unstable_cssModules` option. CSS Modules is now supported automatically.
+Remove `unstable_tailwind` option. Tailwind is now supported automatically.
