+++ packages/remix-dev/compiler/utils/postcss.ts
@@ -19,10 +19,6 @@ const defaultContext: RemixPostcssContext = {
   vanillaExtract: false,
 };
 
-function isPostcssEnabled(config: RemixConfig) {
-  return config.future.unstable_postcss || config.future.unstable_tailwind;
-}
-
 function getCacheKey({ config, context }: Required<Options>) {
   return [config.rootDirectory, context.vanillaExtract].join("|");
 }
@@ -32,10 +28,6 @@ export async function loadPostcssPlugins({
   config,
   context = defaultContext,
 }: Options): Promise<Array<AcceptedPlugin>> {
-  if (!isPostcssEnabled(config)) {
-    return [];
-  }
-
   let { rootDirectory } = config;
   let cacheKey = getCacheKey({ config, context });
   let cachedPlugins = pluginsCache.get(cacheKey);
@@ -45,27 +37,23 @@ export async function loadPostcssPlugins({
 
   let plugins: Array<AcceptedPlugin> = [];
 
-  if (config.future.unstable_postcss) {
-    try {
-      let postcssConfig = await loadConfig(
-        // We're nesting our custom context values in a "remix"
-        // namespace to avoid clashing with other tools.
-        // @ts-expect-error Custom context values aren't type safe.
-        { remix: context },
-        rootDirectory
-      );
+  try {
+    let postcssConfig = await loadConfig(
+      // We're nesting our custom context values in a "remix"
+      // namespace to avoid clashing with other tools.
+      // @ts-expect-error Custom context values aren't type safe.
+      { remix: context },
+      rootDirectory
+    );
 
-      plugins.push(...postcssConfig.plugins);
-    } catch (err) {
-      // If they don't have a PostCSS config, just ignore it.
-    }
+    plugins.push(...postcssConfig.plugins);
+  } catch (err) {
+    // If they don't have a PostCSS config, just ignore it.
   }
 
-  if (config.future.unstable_tailwind) {
-    let tailwindPlugin = await loadTailwindPlugin(config);
-    if (tailwindPlugin && !hasTailwindPlugin(plugins)) {
-      plugins.push(tailwindPlugin);
-    }
+  let tailwindPlugin = await loadTailwindPlugin(config);
+  if (tailwindPlugin && !hasTailwindPlugin(plugins)) {
+    plugins.push(tailwindPlugin);
   }
 
   pluginsCache.set(cacheKey, plugins);
@@ -77,10 +65,6 @@ export async function getPostcssProcessor({
   config,
   context = defaultContext,
 }: Options): Promise<Processor | null> {
-  if (!isPostcssEnabled(config)) {
-    return null;
-  }
-
   let cacheKey = getCacheKey({ config, context });
   let cachedProcessor = processorCache.get(cacheKey);
   if (cachedProcessor !== undefined) {
@@ -105,10 +89,6 @@ let tailwindPluginCache = new Map<string, AcceptedPlugin | null>();
 async function loadTailwindPlugin(
   config: RemixConfig
 ): Promise<AcceptedPlugin | null> {
-  if (!config.future.unstable_tailwind) {
-    return null;
-  }
-
   let { rootDirectory } = config;
   let cacheKey = rootDirectory;
   let cachedTailwindPlugin = tailwindPluginCache.get(cacheKey);
