+++ packages/remix-dev/config.ts
@@ -34,16 +34,8 @@ type Dev = {
   rebuildPollIntervalMs?: number;
 };
 
-export type VanillaExtractOptions = {
-  cache?: boolean;
-};
-
 interface FutureConfig {
-  unstable_cssSideEffectImports: boolean;
   unstable_dev: boolean | Dev;
-  unstable_postcss: boolean;
-  unstable_tailwind: boolean;
-  unstable_vanillaExtract: boolean | VanillaExtractOptions;
 }
 
 /**
@@ -579,12 +571,7 @@ export async function readConfig(
   }
 
   let future: FutureConfig = {
-    unstable_cssSideEffectImports:
-      appConfig.future?.unstable_cssSideEffectImports === true,
     unstable_dev: appConfig.future?.unstable_dev ?? false,
-    unstable_postcss: appConfig.future?.unstable_postcss === true,
-    unstable_tailwind: appConfig.future?.unstable_tailwind === true,
-    unstable_vanillaExtract: appConfig.future?.unstable_vanillaExtract ?? false,
   };
 
   return {
